# specs/001-compensation-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Maliev Compensation Service API
  description: |
    Employee Compensation Management Service API for managing compensation records,
    salary history, benefits enrollment, and bulk operations.

    **Security**: All endpoints require JWT authentication and permission-based authorization.
    All operations are marked as critical and generate audit log entries.
  version: 1.0.0
  contact:
    name: MALIEV Core Developers
    email: support@maliev.com
servers:
  - url: https://api.maliev.com/compensation/v1
    description: Production
  - url: https://staging-api.maliev.com/compensation/v1
    description: Staging
  - url: http://localhost:8080/compensation/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /employees/{employeeId}/compensation:
    get:
      tags: [Compensation]
      summary: Get current compensation details
      description: Retrieves the current compensation record for an employee
      operationId: getCompensationDetails
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
      responses:
        '200':
          description: Compensation details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationRecordDto'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.read]

    post:
      tags: [Compensation]
      summary: Record compensation change
      description: Creates a new compensation record (marks previous as not current)
      operationId: recordCompensationChange
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCompensationChangeDto'
      responses:
        '201':
          description: Compensation change recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: Compensation record created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - BearerAuth: [compensation.update]
  /employees/{employeeId}/compensation/history:
    get:
      tags: [Compensation]
      summary: Get compensation history
      description: Retrieves complete compensation history for an employee
      operationId: getCompensationHistory
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter history from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter history to this date
      responses:
        '200':
          description: Compensation history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalaryHistoryDto'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.read]

  # Benefits Endpoints
...
  /employees/{employeeId}/benefits/{benefitId}/enroll:
    post:
      tags: [Benefits]
      summary: Enroll in benefit
      description: Enrolls an employee in a benefit program
      operationId: enrollInBenefit
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coverageLevel]
              properties:
                coverageLevel:
                  type: string
                  enum: [Individual, Family, Spouse, Spouse+Children]
                employeeContribution:
                  type: number
                  format: decimal
                dependents:
                  type: array
                  items:
                    $ref: '#/components/schemas/DependentDto'
      responses:
        '201':
          description: Enrollment created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - BearerAuth: [compensation.update]
...
  /bulk/salary-increase:
    post:
      tags: [Bulk Operations]
      summary: Apply bulk salary increase
      description: Applies salary increase to multiple employees (max 500 per job)
      operationId: bulkSalaryIncrease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkSalaryIncreaseDto'
      responses:
        '202':
          description: Bulk operation accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJobStatusDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [compensation.admin]
...
  /reports/analysis:
    get:
      tags: [Reports]
      summary: Compensation analysis report
      description: Generates statistical analysis of compensation distribution
      operationId: getCompensationAnalysis
      parameters:
        - name: departmentId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by department
      responses:
        '200':
          description: Analysis report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationAnalysisDto'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [compensation.reports]
  /reports/budget:
    get:
      tags: [Reports]
      summary: Compensation budget report
      description: Calculates total compensation costs and projections
      operationId: getCompensationBudget
      responses:
        '200':
          description: Budget report generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCompensationCost:
                    type: number
                    format: decimal
                  projectedIncrease:
                    type: number
                    format: decimal
                  budgetImpact:
                    type: number
                    format: decimal
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [compensation.reports]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EmployeeIdPath:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Employee unique identifier
  schemas:
    CompensationRecordDto:
      type: object
      required: [id, employeeId, effectiveDate, baseSalary, currency, compensationType]
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        effectiveDate:
          type: string
          format: date-time
        baseSalary:
          type: number
          format: decimal
          description: "[SENSITIVE - Not logged]"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        compensationType:
          $ref: '#/components/schemas/CompensationType'
        bonusPercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        commissionRate:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        changeReason:
          type: string
          maxLength: 500
        approvedBy:
          type: string
          format: uuid
    RecordCompensationChangeDto:
      type: object
      required: [effectiveDate, baseSalary, currency, compensationType, changeReason]
      properties:
        effectiveDate:
          type: string
          format: date-time
          description: Must be today or future date
        baseSalary:
          type: number
          format: decimal
          minimum: 0.01
        currency:
          type: string
          minLength: 3
          maxLength: 3
        compensationType:
          $ref: '#/components/schemas/CompensationType'
        bonusPercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        commissionRate:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        changeReason:
          type: string
          maxLength: 500
    SalaryHistoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        previousSalary:
          type: number
          format: decimal
        newSalary:
          type: number
          format: decimal
        changeAmount:
          type: number
          format: decimal
        changePercentage:
          type: number
          format: decimal
        effectiveDate:
          type: string
          format: date-time
        changeType:
          type: string
          enum: [Promotion, Merit, Adjustment, Initial, Correction]
        changedBy:
          type: string
          format: uuid
...
    DependentDto:
      type: object
      required: [firstName, lastName, relationship, dateOfBirth]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        relationship:
          $ref: '#/components/schemas/DependentRelationship'
        dateOfBirth:
          type: string
          format: date
        nationalId:
          type: string
          description: "[SENSITIVE - Encrypted]"
    BulkSalaryIncreaseDto:
      type: object
      required: [departmentId, increasePercentage, effectiveDate, reason]
      properties:
        departmentId:
          type: string
          format: uuid
        increasePercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        effectiveDate:
          type: string
          format: date-time
        reason:
          type: string
          maxLength: 500
        webhookUrl:
          type: string
          format: uri
          description: Optional callback URL for job completion
    BulkJobStatusDto:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Pending, InProgress, Completed, Failed, PartiallyCompleted]
        affectedEmployees:
          type: integer
        successCount:
          type: integer
        failureCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    CompensationAnalysisDto:
      type: object
      properties:
        averageSalary:
          type: number
          format: decimal
        medianSalary:
          type: number
          format: decimal
        minSalary:
          type: number
          format: decimal
        maxSalary:
          type: number
          format: decimal
        totalEmployees:
          type: integer
        byCompensationType:
          type: object
          additionalProperties:
            type: integer
        byDepartment:
          type: object
          additionalProperties:
            type: number
            format: decimal

    # Enums
    CompensationType:
      type: string
      enum: [Salary, Hourly, Contract, Commission]
...
    DependentRelationship:
      type: string
      enum: [Spouse, Child, DomesticPartner, Parent, Other]

    # Error Responses
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidSalary:
              value:
                code: INVALID_SALARY
                message: Salary must be positive
            invalidCurrency:
              value:
                code: INVALID_CURRENCY
                message: Invalid currency code
            invalidEffectiveDate:
              value:
                code: INVALID_EFFECTIVE_DATE
                message: Date must be current or future
            changeLimitExceeded:
              value:
                code: CHANGE_LIMIT_EXCEEDED
                message: Increase exceeds 25% limit
    Forbidden:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_AUTHORIZED
            message: User cannot access compensation
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            employeeNotFound:
              value:
                code: EMPLOYEE_NOT_FOUND
                message: Employee not found
            benefitNotFound:
              value:
                code: BENEFIT_NOT_FOUND
                message: Benefit not found
    Conflict:
      description: Conflict - duplicate or concurrent update
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            enrollmentExists:
              value:
                code: ENROLLMENT_EXISTS
                message: Already enrolled in benefit
            concurrentUpdate:
              value:
                code: CONCURRENT_UPDATE
                message: Record was modified by another user, please refresh and retry